+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+v0.0.5
+- Added option to create a local backup of the files
+- Added check_packages to check that all required packages are present
+- Deleting files now requires to reenter the key
+- Added linux and mac support (Not tested)
+- Minor changes
+
 v0.0.4
 - Added the move records function
 - Added changelog